product-catalog/requirements.txt
azure-functions
mysql-connector-python==8.2.0
azure-storage-blob
